Exclusão de Produto

+++ backend/storecheckoutsystem/src/main/java/com/example/storecheckoutsystem/repository/ProdutoRepository.java
@@ -9,6 +9,9 @@ import com.example.storecheckoutsystem.model.Produto;
 
 
 @Repository
-public interface ProdutoRepository extends JpaRepository<Produto, Long>{
+public interface ProdutoRepository extends JpaRepository<Produto, Integer>{
+
+    
+
     
 }
